loans/api_loans/api/serializers.py
from rest_framework import serializers
from api.models import Emprestimo

class EmprestimoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Emprestimo
        fields = '__all__'