pyproject.toml
[tool.poetry]
name = "loan_api"
version = "0.1.0"
description = ""
authors = ["Arthur Antunes <arthur.antunes.aa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
